Fixed up Node tests after math folder move
